feat(104183): Valida data inicio e encerramento da conta no retorno de saldos bancários
